feat(react): update lineage sidebar ui/ux
